pyproject.toml
[tool.poetry]
name = "AICallsAnalytics"
version = "0.1.0"
description = "AICallsAnalytics"
authors = ["alex <author@domain>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.2"}
sqlalchemy = "^2.0.36"
alembic = "^1.13.3"
pydantic-settings = "^2.6.0"
pytest = "^8.3.3"
loguru = "^0.7.2"
black = "^24.10.0"
mypy = "^1.12.1"
ruff = "^0.7.0"
uvicorn = "^0.32.0"
pyhumps = "^3.8.0"
redis = "^5.1.1"
fakeredis = "^2.25.1"
pytest-asyncio = "^0.24.0"
asyncpg = "^0.30.0"
orjson = "^3.10.9"
gunicorn = "^23.0.0"
alembic-postgresql-enum = "^1.4.0"
aiocache = "^0.12.3"
aiohttp = "^3.12.15"
aiormq = "^6.9.0"
faststream = {extras = ["rabbit"], version = "^0.5.48"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"